Extract inline tests from 14 remaining component files
Convert 13 single-file components (button, checkbox, input_field,
key_hints, menu, progress_bar, radio_group, router, select,
selectable_list, spinner, tabs, focus_manager) from foo.rs to
foo/mod.rs + foo/tests.rs directory modules.

Extract component/mod.rs tests to sibling component/tests.rs.

This brings all component test files to a consistent pattern
matching the accordion/breadcrumb/dialog style established in
prior refactoring.
